% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{ClassificationCurves}
\alias{ClassificationCurves}
\title{ClassificationCurves}
\usage{
ClassificationCurves(
  data,
  CONNECTORDataClustered,
  cores = 1,
  entropyCutoff = 1,
  probCutoff = 0.6
)
}
\arguments{
\item{data}{CONNECTORData. See CONNECTORData for details.}

\item{CONNECTORDataClustered}{Configuration choosen by ConfigSelection()}

\item{cores}{number of cores used}

\item{entropyCutoff}{...}

\item{probCutoff}{...}
}
\value{
...
}
\description{
...
}
\seealso{
CONNECTORDataClustered()
}
