% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCluster.R
\name{ClusterPlot}
\alias{ClusterPlot}
\title{ClusterPlot}
\usage{
ClusterPlot(CONNECTORDataClustered, feature = NULL, feature_type = "auto")
}
\arguments{
\item{CONNECTORDataClustered}{Connector object created with ConnectorData}

\item{feature}{dunno}

\item{feature_type}{Character string specifying how to treat the feature: "auto" (default), "discrete", or "continuous"}
}
\value{
...
}
\description{
plot ??
}
\seealso{
CONNECTORDataClustered ConfigSelection ClusterAnalysis
}
