% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configSelection.R
\name{configSelection}
\alias{configSelection}
\title{configSelection}
\usage{
configSelection(results, G, best)
}
\arguments{
\item{results}{ClusterAnalysis output}

\item{G}{Number of cluster selected}

\item{best}{Setup choosen. "MaxFreq" or "MinfDB"}
}
\value{
an object with the cofiguration selected
}
\description{
Print fDB and TT violin plots with best and most frequent value
}
\seealso{
ClusterAnalysis
}
