% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORData-class.R
\docType{class}
\name{CONNECTORData-class}
\alias{CONNECTORData-class}
\title{CONNECTORData}
\description{
CONNECTORData is a S4 objectwith four arguments: (i) curves: tibble with n columns (i.e. ID, time and n-2 columns of records) encoding the growth data for each sample, (ii) length: the tibble reporting the number of observations collected per sample  (iii) annotation: the tibble matching the samples with their annotations, (iv) TimeGrid: the vector storing all the sample time points (i.e. time grid). Furthermore, it prints a brief summary of the input data, i.e. the total number of curves (samples), the minimum and the maximum curve length.
}
\section{Slots}{

\describe{
\item{\code{curves}}{tibble with n columns (i.e. ID, time and n-2 columns of records) encoding the growth data for each sample}

\item{\code{dimensions}}{the tibble reporting the number of observations collected per sample}

\item{\code{annotations}}{the tibble matching the samples with their annotations}

\item{\code{TimeGrids}}{the vector storing all the sample time points (i.e. time grid)}
}}

