% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\docType{class}
\name{CONNECTORDataClassified-class}
\alias{CONNECTORDataClassified-class}
\title{ClassificationCurves}
\arguments{
\item{data}{CONNECTORData. See CONNECTORData for details.}

\item{CONNECTORDataClustered}{Configuration choosen by ConfigSelection()}

\item{cores}{number of cores used}

\item{entropyCutoff}{...}

\item{probCutoff}{...}
}
\value{
...
}
\description{
...

Objects of this class hold the results of classifying new curves into existing clusters.
}
\section{Slots}{

\describe{
\item{\code{ClassMatrix}}{Data frame with classification probabilities for each subject.}

\item{\code{ClassMatrix_entropy}}{Data frame with classification results including entropy and certainty thresholds.}

\item{\code{ClassificationData}}{The CONNECTORData object containing the new curves.}

\item{\code{ClusteredData}}{The CONNECTORDataClustered model used for classification.}
}}

\seealso{
CONNECTORDataClustered()
}
