% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORDataClustered-class.R
\docType{class}
\name{CONNECTORDataClustered-class}
\alias{CONNECTORDataClustered-class}
\title{CONNECTORDataClustered}
\description{
CONNECTORDataClustered is a class with the set choosen and with KData that is a list with all output of PRESETKMEANS that are necessary for call JUSTKMEANS and INTFCLUST
}
\section{Slots}{

\describe{
\item{\code{TTandfDBandSil}}{Contains TT, fDB, Silhouette and G choosen}

\item{\code{CfitandParameteres}}{Contains values calculated during clusteranalysis}

\item{\code{h}}{h value choosen}

\item{\code{freq}}{frequency of the clusterization}

\item{\code{cluster.names}}{names of the clusters}

\item{\code{KData}}{contains: CData Contains timeseriefile datas, TimeGrids TimeGrids present in CONNECTORData, points Initial coefficent spline points, N subject number, S block diagonal matrix, FullS full block diagonal matrix}
}}

