% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{fclust.curvepred}
\alias{fclust.curvepred}
\title{fclust.curvepred}
\usage{
fclust.curvepred(data, tau = 0.95, tau1 = 0.975, q, KData)
}
\arguments{
\item{data}{...}

\item{tau}{...}

\item{tau1}{...}

\item{q}{...}

\item{KData}{...}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
