% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{fclust}
\alias{fclust}
\title{fclust}
\usage{
fclust(
  data_input,
  q,
  h,
  K,
  pert = 0.01,
  pert1 = 0.01,
  natural = TRUE,
  tol = 0.001,
  maxit = 20,
  hard = FALSE,
  pert2 = 0,
  seed = NULL
)
}
\arguments{
\item{data_input}{}

\item{q}{}

\item{h}{}

\item{K}{}

\item{pert}{}

\item{pert1}{}

\item{natural}{}

\item{tol}{}

\item{maxit}{}

\item{hard}{}

\item{pert2}{}

\item{seed}{}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
