% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{fclust_pred}
\alias{fclust_pred}
\title{fclust_pred}
\usage{
fclust_pred(fit, data = NULL, reweight = FALSE, pert2 = 0)
}
\arguments{
\item{fit}{...}

\item{data}{....}

\item{pert2}{...}

\item{rewight}{...}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
