% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{fclustconst}
\alias{fclustconst}
\title{fclustconst}
\usage{
fclustconst(parameters, vars, S)
}
\arguments{
\item{parameters}{...}

\item{vars}{...}

\item{S}{...}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
