% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{fclustEstep}
\alias{fclustEstep}
\title{fclustEstep}
\usage{
fclustEstep(parameters, curve_ok, vars, S, hard)
}
\arguments{
\item{parameters}{...}

\item{curve_ok}{...}

\item{vars}{...}

\item{S}{....}

\item{hard}{...}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
