% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{fclustMstep}
\alias{fclustMstep}
\title{fclustMstep}
\usage{
fclustMstep(parameters, curve_ok, S, vars, hard, p, pert1, tol)
}
\arguments{
\item{parameters}{...}

\item{curve_ok}{...}

\item{S}{....}

\item{vars}{...}

\item{hard}{...}

\item{p}{...}

\item{pert1}{...}

\item{tol}{...}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
