% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\alias{filterData,CONNECTORData-method}
\title{filterData}
\usage{
filterData(data, minPoints = 3)

\S4method{filterData}{CONNECTORData}(data, minPoints = 3)
}
\arguments{
\item{data}{CONNECTORData. (see \code{\link{ConnectorData}})}

\item{minPoints}{Minimum number of points required for each measure. If a subject has fewer than this value in any of its measurements, it will be removed. Default is 3.}
}
\value{
Returns a new CONNECTORData object with the filtered subjects.
}
\description{
Filters the CONNECTORData object by removing subjects that have less than a certain number of points in at least one of their measurements.
}
