% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORDataClustered-class.R
\name{getClusterNames}
\alias{getClusterNames}
\alias{getClusterNames,CONNECTORDataClustered-method}
\title{getClusterNames}
\usage{
getClusterNames(object)

\S4method{getClusterNames}{CONNECTORDataClustered}(object)
}
\arguments{
\item{object}{CONNECTORDataClustered object}
}
\value{
Character vector of cluster names, or NULL if no names have been set.
}
\description{
Get the custom names assigned to clusters in a CONNECTORDataClustered object.
}
\examples{
\dontrun{
# Get cluster names
getClusterNames(clustered_data)
}
}
