% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORDataClustered-class.R
\name{getMeasures}
\alias{getMeasures}
\title{getMeasures}
\usage{
getMeasures(object)
}
\arguments{
\item{object}{CONNECTORData or CONNECTORDataClustered object}
}
\value{
A vector of measure names.
}
\description{
Extract and display measures from CONNECTORData or CONNECTORDataClustered object.
}
