% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORDataClustered-class.R
\name{getMembershipProbs}
\alias{getMembershipProbs}
\alias{getMembershipProbs,CONNECTORDataClustered-method}
\title{getMembershipProbs}
\usage{
getMembershipProbs(object)

\S4method{getMembershipProbs}{CONNECTORDataClustered}(object)
}
\arguments{
\item{object}{CONNECTORDataClustered object}
}
\value{
A tibble with subjID and membership probabilities for each cluster.
}
\description{
Get the cluster membership probabilities for each subject.
}
\examples{
\dontrun{
# Get membership probabilities
getMembershipProbs(clustered_data)
}
}
