% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORDataClustered-class.R
\name{getParameters}
\alias{getParameters}
\alias{getParameters,CONNECTORDataClustered-method}
\title{getParameters}
\usage{
getParameters(object)

\S4method{getParameters}{CONNECTORDataClustered}(object)
}
\arguments{
\item{object}{CONNECTORDataClustered object}
}
\value{
A list containing G, p, and h.
}
\description{
Get the parameters (G, p, h) used for the clustering in a CONNECTORDataClustered object.
}
\examples{
\dontrun{
# Get parameters
getParameters(clustered_data)
}
}
