% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexPlotExtrapolation.R
\name{IndexPlotExtrapolation}
\alias{IndexPlotExtrapolation}
\title{IndexPlotExtrapolation}
\usage{
IndexPlotExtrapolation(results)
}
\arguments{
\item{results}{ClusterAnalysis output}
}
\value{
fDB and TT violin plots with best and most frequent value
}
\description{
Print fDB and TT violin plots with best and most frequent value
}
\seealso{
ClusterAnalysis
}
