% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexPlotExtrapolation2.R
\name{IndexPlotExtrapolation2}
\alias{IndexPlotExtrapolation2}
\title{IndexPlotExtrapolation2}
\usage{
IndexPlotExtrapolation2(CONNECTORData, ConfigChosen, feature)
}
\arguments{
\item{CONNECTORData}{Connector object created with ConnectorData}

\item{ConfigChosen}{configuration choosen using ConfigSelection}

\item{feature}{dunno}
}
\value{
...
}
\description{
plot ??
}
\seealso{
ConnectorData ConfigSelection ClusterAnalysis
}
