% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORData-methods.R
\name{initialize,CONNECTORData-method}
\alias{initialize,CONNECTORData-method}
\title{ConnectorData}
\usage{
\S4method{initialize}{CONNECTORData}(.Object, curves, dimensions, annotations, TimeGrids)
}
\arguments{
\item{TimeSeriesFile}{The name of the excel/csv file storing the  growth evolution data. Alternatively accepts a tibble with the data.}

\item{AnnotationFile}{The name of a excel/csv file storing  the annotation data. Alternatively accepts a tibble with the annotations.}
}
\value{
ConnectorData returns a S4 object, called ConnectorData, with four arguments: (i) curves: tibble with 5 columns (subjID, measureID, time, curvesID, value) encoding the growth data for each sample, (ii) dimensions: the tibble reporting the number of observations collected per sample  (iii) annotations: the tibble matching the samples with their annotations, (iv) TimeGrid: the vector storing all the sample time points (i.e. time grid). Furthermore, it prints a brief summary of the input data, i.e. the total number of curves (samples), the minimum and the maximum curve length.
}
\description{
ConnectorData is a function that imports the data from the excel file storing the growth evolution data and the annotation file.
}
\details{
create CONNECTORData like an object that contains 4 different slots curves dimensions annotations TimeGrid
curves is a tibble that contains subjID, measureID, time, curvesID (concatenation of measureID_subjID), value
dimensions is a tibble that contains the number of observations for each ID
annotations is a tibble that contains the ID of each curve
TimeGrid is a vector that contains the time points of the curves
}
\examples{
TimeSeriesFile<-system.file("testdata", "test.xlsx", package = "ConnectorV2.0")
AnnotationFile <-system.file("testdata", "testinfo.txt", package = "ConnectorV2.0")
}
