% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{intfclust}
\alias{intfclust}
\title{intfclust}
\usage{
intfclust(
  q,
  h,
  pert = 0,
  K,
  class,
  CLUSTData,
  pert1 = 0.01,
  tol = 0.001,
  maxit = 20,
  hard = FALSE,
  pert2 = 0
)
}
\arguments{
\item{q}{...}

\item{h}{...}

\item{pert}{...}

\item{K}{...}

\item{class}{...}

\item{CLUSTData}{...}

\item{pert1}{...}

\item{tol}{...}

\item{maxit}{...}

\item{hard}{.....}

\item{pert2}{...}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
