% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{justKmeans}
\alias{justKmeans}
\title{justKmeans}
\usage{
justKmeans(CLUSTData, K)
}
\arguments{
\item{CLUSTData}{data}

\item{k}{number of clusters}
}
\value{
results of kmeans
}
\description{
uses kmeans
}
\seealso{
kmeans()
}
