% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimes.R
\name{plotTimes}
\alias{plotTimes}
\title{plotTimes}
\usage{
plotTimes(data, large = FALSE, measureIDs = NULL)
}
\arguments{
\item{data}{CONNECTORData. (see \code{\link{DataImport}})}

\item{large}{logical. If TRUE the function will return a plot that represent time grid density.}
}
\value{
Data Visualization returns a plot with the density time grid and the line plot of growth data as a ggplot object.
In details, a point $p_{x,y}$ of the time grid density  is defined by a pair of coordinates $p_{x,y}=( x,y) \ $ and by a colour. $p_{x,y}$ is defined if only if exists at least one sample with two observations at time $x\ $ and $y$.
The colour associates with it encodes the frequency of samples in which $p_{x,y}$ is present.
}
\description{
creates a barplot of the time presences in various IDs or also a plot that represent time grid density.
}
