% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTimeSeries.R
\name{PlotTimeSeries}
\alias{PlotTimeSeries}
\title{PlotTimeSeries}
\usage{
PlotTimeSeries(data, feature = NULL, labels)
}
\arguments{
\item{data}{CONNECTORData. See CONNECTORData for details.}

\item{feature}{The column name reported in the AnnotationFile containing the feature to be investigated.}

\item{labels}{The vector containing the text for the axis names and plot title.}
}
\value{
PlotTimeSeries returns a ggplot object.
}
\description{
the line plot of the time series data. The curves are colored with respect to the feature chosen by the user and/or reported in the AnnotationFile.
}
