% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust.R
\name{presetKmeans}
\alias{presetKmeans}
\title{presetKmeans}
\usage{
presetKmeans(CData, natural = TRUE, q, pert = 0.01)
}
\arguments{
\item{CData}{data}

\item{natural}{...}

\item{q}{...}

\item{pert}{perturbation}
}
\value{
a list with all necessary data
}
\description{
presets data for the kmeans
}
\seealso{
kmeans()
}
