% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{printReportSummary}
\alias{printReportSummary}
\title{Print method for analysis report}
\usage{
printReportSummary(report)
}
\arguments{
\item{report}{Report object generated by generateReport}
}
\description{
Print method for analysis report
}
