% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORDataClustered-class.R
\name{setClusterNames}
\alias{setClusterNames}
\alias{setClusterNames,CONNECTORDataClustered-method}
\title{setClusterNames}
\usage{
setClusterNames(object, names)

\S4method{setClusterNames}{CONNECTORDataClustered}(object, names)
}
\arguments{
\item{object}{CONNECTORDataClustered object}

\item{names}{Character vector of cluster names. Length must match the number of clusters.}
}
\value{
Updated CONNECTORDataClustered object with cluster names set.
}
\description{
Set custom names for clusters in a CONNECTORDataClustered object.
These names will be used in plots and dataframes.
}
\details{
This method allows you to assign meaningful names to clusters instead of using
numeric identifiers. The names are stored in the object and used by other methods
like getClusters, getClustersCentroids, and clusterDistribution.
}
\examples{
\dontrun{
# Set names for a 3-cluster solution
clustered_data <- setClusterNames(clustered_data, c("Low", "Medium", "High"))

# Now getClusters will return named clusters
getClusters(clustered_data)
}
}
