% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SilhouetteAndEntropy.R
\name{SilEntropy}
\alias{SilEntropy}
\title{SilhouetteAndEntropy}
\usage{
SilEntropy(ConfigChosen)
}
\arguments{
\item{ConfigChosen}{set choosen with configSelection()}
}
\value{
a plot.
}
\description{
plot a graph with silhouette and entropy
}
