% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinePlot.R
\name{splinePlot}
\alias{splinePlot}
\alias{splinePlot,CONNECTORDataClustered-method}
\title{splinePlot}
\usage{
splinePlot(CONNECTORDataClustered)

\S4method{splinePlot}{CONNECTORDataClustered}(CONNECTORDataClustered)
}
\arguments{
\item{CONNECTORDataClustered}{data created with selectCluster()}
}
\value{
a plot of splines for each data
}
\description{
plot splines
}
