% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{truncate}
\alias{truncate}
\title{truncate}
\usage{
truncate(data, feature = NULL, truncTime = NULL, labels = NULL, measure = NULL)
}
\arguments{
\item{data}{CONNECTORData. (see \code{\link{ConnectorData}})}

\item{truncTime}{A two dimension vector of integers corresponding to the time points where the curves will be truncated. If an integer number is passed, than it will be considered as the upper time point by default.}

\item{measure}{Measure(s) on which to perform the Truncation. If NULL (default), all measures are truncated.}
}
\value{
truncate returns a list containing the truncated data and the plot of the truncated curves.
}
\description{
Truncates the functional data (the time series) at a specific time point chosen by the user.
}
