% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncatePlot.R
\name{truncatePlot}
\alias{truncatePlot}
\alias{truncatePlot,CONNECTORData-method}
\title{truncatePlot}
\usage{
truncatePlot(
  data,
  feature = NULL,
  truncTime = NULL,
  labels = NULL,
  measure = NULL
)

\S4method{truncatePlot}{CONNECTORData}(
  data,
  feature = NULL,
  truncTime = NULL,
  labels = NULL,
  measure = NULL
)
}
\arguments{
\item{data}{CONNECTORData. (see \code{\link{ConnectorData}})}

\item{feature}{The column name reported in the AnnotationFile containing the feature  to be investigated.}

\item{truncTime}{A two dimension vector of integers corresponding to the time points where the curves will be truncated. If an integer number is passed, than it will be considered as the upper time point by default.}

\item{labels}{Vector containing the text for the title of axis names and plot title.}

\item{measure}{Measure on which to perform the Truncation.}
}
\value{
a plot with a bar that indicate the position of truncation. Doesen0t truncate effectively. For truncation check truncate()
}
\description{
Plot the truncation of functional data (the time series) at a specific time point chosen by the user.
}
