% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCluster.R
\name{ClusterPlot}
\alias{ClusterPlot}
\title{ClusterPlot}
\usage{
ClusterPlot(CONNECTORDataClustered, feature = NULL)
}
\arguments{
\item{CONNECTORDataClustered}{Connector object created with ConnectorData}

\item{feature}{dunno}
}
\value{
...
}
\description{
plot ??
}
\seealso{
CONNECTORDataClustered ConfigSelection ClusterAnalysis
}
