% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscriminantPlot.R
\name{DiscriminantPlot}
\alias{DiscriminantPlot}
\title{DiscrimintantPlot}
\usage{
DiscriminantPlot(CONNECTORDataClustered, feature)
}
\arguments{
\item{CONNECTORDataClustered}{data created with selectCluster()}

\item{feature}{feature selected for coloring}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
