% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscriminantPlot.R
\name{DiscriminantPlot}
\alias{DiscriminantPlot}
\title{DiscrimintantPlot}
\usage{
DiscriminantPlot(CONNECTORData, ConfigChosen, feature)
}
\arguments{
\item{CONNECTORData}{data created by ConnectorData()}

\item{ConfigChosen}{data created with configSelection()}

\item{feature}{feature selected for coloring}
}
\value{
...
}
\description{
...
}
\seealso{
...
}
