% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCluster.R
\name{estimateCluster}
\alias{estimateCluster}
\alias{estimateCluster,CONNECTORData-method}
\title{estimateCluster}
\usage{
estimateCluster(
  CONNECTORData,
  G,
  p,
  h = NULL,
  runs = 50,
  seed = 2404,
  cores = 1,
  PercPCA = 0.85,
  MinErrFreq = 0,
  pert = 0.01
)

\S4method{estimateCluster}{CONNECTORData}(
  CONNECTORData,
  G,
  p,
  h = NULL,
  runs = 50,
  seed = 2404,
  cores = 1,
  PercPCA = 0.85,
  MinErrFreq = 0,
  pert = 0.01
)
}
\arguments{
\item{CONNECTORData}{CONNECTORList. (see \code{\link{ConnectorData}})}

\item{G}{The vector/number of possible clusters.}

\item{p}{The dimension of the natural cubic spline basis. (see \code{\link{BasisDimension.Choice}})}

\item{h}{...}

\item{runs}{Number of runs.}

\item{seed}{Seed for the kmeans function.}

\item{cores}{Number of cores to parallelize computations.}

\item{PercPCA}{... standar 0.85}

\item{MinErrFreq}{...}

\item{pert}{....}
}
\value{
StabilityAnalysis returns a list of (i) lists, called ConsensusInfo, reporting for each G and h: the Consensus Matrix, either as a NxN matrix, where N is the number of samples, or plot, and the most probable clustering obtained from running several times the method; (ii) the box plots showing both the Elbow plot considering the total tightness and the box plots of the fDB indexes for each G; and finally, (iii) the seed. See \code{\link{IndexesPlot.Extrapolation}} and \code{\link{MostProbableClustering.Extrapolation}}.
}
\description{
Fits and clusters the data with respect to the Functional Clustering Model \link{Sugar and James}. Multiple runs of the algorithm are necessary since the algorithm is stochastic As explained in \link{Sugar and James}, to have a simple low-dimensional representation of the individual curves and to reduce the number of parameters to be estimated, h value must be equals or lower than \eqn{min(p,G-1)}.
}
\details{
...
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
