% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatepDimension.R
\name{estimatepDimension}
\alias{estimatepDimension}
\title{estimatepDimension}
\usage{
estimatepDimension(data, i = NULL, p, cores = 1)
}
\arguments{
\item{data}{CONNECTORData. See CONNECTORData for details.}

\item{i}{the measure name to be used for the analysis. For example, if i="measure1" the first measure of the data set will be used for the analysis. If i=c("measure1","measure2") the first and the second measures of the data set will be used for the analysis. Default value is NULL.}

\item{p}{The vector of the dimensions of the natural cubic spline basis}

\item{cores}{The number of Cores to be used for parallel computation. Maximum number is 10.}
}
\value{
Returns a list containing for each element a line plot of the cross-validated loglikelihood for each value of p, in grey the result of all ten repetitions of the likelihood calculation and in black the mean of them.
}
\description{
Generates a line plot reporting the cross-validated loglikelihood value for each number of knots. In details, for each number of knots 10\% of the curves from the whole data are removed and treated as a test set, then the remaing curves are fitted using the FCM and the loglikelihood on the test set is calculated. The process is then repeated nine more times.
}
