% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORDataClustered-class.R
\name{getAnnotations}
\alias{getAnnotations}
\alias{getAnnotations,CONNECTORDataClustered-method}
\title{getAnnotations}
\usage{
getAnnotations(object)

\S4method{getAnnotations}{CONNECTORDataClustered}(object)
}
\arguments{
\item{object}{CONNECTORDataClustered object}
}
\value{
The annotations tibble with cluster assignments
}
\description{
Extract and display annotations from CONNECTORDataClustered object
}
