% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONNECTORDataClustered-class.R
\name{getClusters}
\alias{getClusters}
\title{getClusters}
\usage{
getClusters(object)
}
\arguments{
\item{object}{CONNECTORDataClustered object}
}
\value{
A dataframe with the cluster association for each subjID.
}
\description{
Extract and display clusters from  CONNECTORDataClustered object.
Shows all available features (annotation columns) in both cases.
}
\details{
This method provides the features available in the annotations of the provided object.
}
\examples{
\dontrun{

# For CONNECTORDataClustered 
getClusters(my_clustered_data)
}
}
