% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumDiscriminationFunction.R
\name{MaximumDiscriminationFunction}
\alias{MaximumDiscriminationFunction}
\title{Maximum Discrimination Function}
\usage{
MaximumDiscriminationFunction(CONNECTORDataClustered, absvalue = TRUE)
}
\arguments{
\item{CONNECTORDataClustered}{The CONNECTORDataClustered object obtained from selectCluster function.}

\item{absvalue}{If TRUE, the absolute values of the weights are plotted.
@return
MaximumDiscriminationFunction generates two plots as ggplot objects, showing
the weights applied to each dimension for determining the cluster membership, plotted as a single curve.}
}
\description{
Visualizes the h curve(s) representing the optimal weights to apply to each dimension for determining the cluster membership.
}
\references{
Gareth M. James and Catherine A. Sugar, (2003). Clustering for Sparsely Sampled Functional Data. Journal of the American Statistical Association.
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
