% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{plot

create  plot of the basic data or the data already analyzed}
\usage{
plot(data, ...)
}
\arguments{
\item{data}{CONNECTORData. (see \code{\link{ConnectorData}})}
}
\description{
plot

create  plot of the basic data or the data already analyzed
}
