% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCluster.R
\name{selectCluster}
\alias{selectCluster}
\title{selectCluster}
\usage{
selectCluster(results, G, best)
}
\arguments{
\item{results}{ClusterAnalysis output}

\item{G}{Number of cluster selected}

\item{best}{Setup choosen. "MaxFreq" or "MinfDB"}
}
\value{
an object with the configuration selected
}
\description{
Select cluster configuration by printing fDB and TT violin plots with best and most frequent value
}
\seealso{
ClusterAnalysis
}
