% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpline.R
\name{splinePlot}
\alias{splinePlot}
\title{splinePlot}
\usage{
splinePlot(CONNECTORDataClustered)
}
\arguments{
\item{CONNECTORDataClustered}{data created with selectCluster()}
}
\value{
a plot of splines for each data
}
\description{
plot splines
}
