% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinePlot.R
\name{splinePlot}
\alias{splinePlot}
\alias{splinePlot,ANY-method}
\title{splinePlot}
\usage{
splinePlot(ConfigChosen)

\S4method{splinePlot}{ANY}(ConfigChosen)
}
\arguments{
\item{ConfigChosen}{data created with configSelection()}
}
\value{
a plot of splines for each data
}
\description{
plot splines
}
