% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SilhouetteAndEntropy.R
\name{validateCluster}
\alias{validateCluster}
\title{validateCluster}
\usage{
validateCluster(CONNECTORDataClustered)
}
\arguments{
\item{CONNECTORDataClustered}{object of class CONNECTORDataClustered created with selectCluster()}
}
\value{
a list containing the combined plot and table with clustering metrics.
}
\description{
Validate clustering results by calculating and plotting silhouette and entropy metrics
}
